feat(query_builder): Allow force removal of unreferenced optional joins
Updates the semantics of the `LeftJoinOptional()` method.

In scenarios where the left table has a one-to-one or one-to-zero
relationship with the joined right table, a user knows that removing
the left join will have no impact on the result if the joined table
is not referenced elsewhere.

This change allows the user to declare this and have the unreferenced
join be removed, thus optimizing the final generated SQL.