feat(scanner): Introduce `default_tables` tag for default table aliases
This commit introduces a new tag key, `default_tables`, to define a default table alias for subfields and subsequent sibling fields.

Previously, this functionality was implicitly tied to anonymous fields, which was restrictive. The new `default_tables` tag can be used on any field, including a dummy field (_ struct{}), to provide a table context. This offers greater flexibility in struct design.

The documentation and examples in example_test.go have been updated to reflect this change.